remove/isolate old unnecessary modules
