--- torchrl/agents/__init__.py
@@ -1,2 +1,0 @@
-from .base_agent import BaseAgent
-from .gym_random_agent import GymRandomAgent
