--- torchrl/contrib/agents/__init__.py
@@ -1,2 +1,0 @@
-from .a2c_agent import BaseA2CAgent
-from .ppo_agent import BasePPOAgent
