--- torchrl/contrib/problems/__init__.py
@@ -1,3 +1,0 @@
-from .a2c import A2CProblem
-from .ppo import PPOProblem
-from .prioritized_dqn import PrioritizedDQNProblem
