--- torchrl/policies/__init__.py
@@ -1,1 +1,0 @@
-from .epsilon_greedy import epsilon_greedy
